Feature: Added navigation capabilities to the BottomNavigationSheet.
Clicking any of the three MenuItems in the BottomNavigationSheet will navigate to the corresponding fragment.
For this functionality I have created a enum class Navigate to indicate to which fragment we should navigate.
I have also added a BindingAdapter in Util.kt to shortcut the NavigationView.setOnNavigationItemSelectedListener and use XML expressions with a lambda to directly call a function in the BottomNavigationSheetViewModel.kt. This lambda is achieved using a interface MenuItemSelected also in Util.kt.